Bug #557
git-svn-id: https://svn.java.net/svn/electric~svn/trunk@3276 b64ed6c1-75a8-e629-b1da-d0dae74a743a

+++ srcj/com/sun/electric/database/ImmutableVariable.java
@@ -237,7 +237,7 @@ public class ImmutableVariable {
             return (byte)(type.byteValue()|ARRAY);
         } else {
             Byte type = (Byte)validClasses.get(value.getClass());
-            return type;
+            return type.byteValue();
         }
     }
